+++ src/components/shared/PlayerHeadshot.jsx
@@ -1,13 +1,11 @@
 import React from 'react';
 
-const PlayerHeadshot = ({ src }) => {
+const PlayerHeadshot = ({ src, className = '' }) => {
   return (
-    <div className="w-[200px] h-[200px] rounded-xl overflow-hidden border-2 border-black">
-      <img
-        src={src}
-        alt="Player headshot"
-        className="w-full h-full object-cover"
-      />
+    <div
+      className={`w-[200px] h-[200px] rounded-xl overflow-hidden border-2 border-black ${className}`}
+    >
+      <img src={src} alt="Player headshot" className="w-full h-full object-cover" />
     </div>
   );
 };
